content: add all blog posts, conversion script
This change adds all blog posts, converted from Nikola. The conversion
script (AWK) is added for reference, in the `scripts/` folder.
Also, a separator space is added to the left hand side menu, between
"About" and the first blog post, so as to visually distinguish the blog
post links from "Github" and "About".

Signed-off-by: Thilo Fromm <thilo.alexander@gmail.com>